Refactored classes to conform to stat sorter class styling
