+++ lib/stat_sorter.rb
@@ -66,7 +66,7 @@ class StatSorter
 
   def regular_away(game)
     team = teams_by_id[game.away_team_id]
-
+    
     team.regular[:goals] += game.away_goals
     team.regular[:games] += 1
     team.regular[:goals_against] += game.home_goals
